[windows] Handle different installation mode
So far, only the MSI installer was supported. With this change (along
with the rest of the Omnibus 4 build), it is now possible to use the
source install script as well (even though we don't have one yet on
Windows). Finally, just git pull-ing the `dd-agent` repo, installing
the requirements and running the agent (YES, even as a regular
Windows service) has also been made possible.

This commit is particularly intented to people willing to get hack on
the agent.